ui(news): show correct loading indicator in ArticlePage
